+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v2.0.0:
 
 - breaking change: omitting `--listen-address` now means that systemd socket activation should be used.
+- breaking change: when no substituter is specified, refuse to run instead of running a server that always returns 404.
 
 v1.0.1:
 - fix flaky test